requirements.txt
anyio==4.10.0
asgiref==3.9.1
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cloudinary==1.44.1
cryptography==45.0.6
Django==5.2.5
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
djangorestframework==3.16.1
firebase_admin==7.1.0
gcloud==0.18.3
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.3.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.74.0
gunicorn==23.0.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
jwcrypto==1.5.6
msgpack==1.1.1
oauth2client==4.1.3
packaging==25.0
proto-plus==1.26.1
protobuf==6.32.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.23.0
PyJWT==2.10.1
pyparsing==3.2.3
Pyrebase4==4.8.0
python-jwt==4.1.0
requests==2.32.5
requests-toolbelt==0.10.1
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
urllib3==1.26.20
whitenoise==6.9.0
